embryo {BioSSA}	R Documentation

Create a new ‘embryo’ object

Description

Creates a new object of class ‘embrio2d’ or ‘embrio3d’ correspondingly, input and output data for BioSSA procedure.

ASH: Probably, we should join these methods into one and determine dimension by formula length.

Usage

## S3 method for class 'formula'
embryo2d(x, data = NULL, topology = c(Inf, Inf), ...)
## S3 method for class 'formula'
embryo3d(x, data = NULL, ...)

Arguments

x
formula which describes 2d or 3d spatial embryo data

...
further arguments passed to model.frame function

data
a data frame (or more precisely, anything that is a valid envir argument in eval, e.g., a list or an environment) containing values for any variables in the formula. If not found in data, or if data is unspecified, the variables are looked for in the environment of the formula

topology
field topology, a vector of two positive (may be Inf) values. Inf means no cycling by corresponding coordinate and finite number denotes circle length

Details

topology argument is used if 2d data given by cylindrical sweeping of 3d data.

Value

Object of class ‘embryo2d’ or ‘embryo3d’ correspondingly.

‘embrio2d’ is list with following fields: two coordinates 'x2d', 'y2d' and gene activity 'value'. List has 'topology' attribute, which stores topology. If object returned from \сodereconstruct or residuals methods of ‘BioSSA2d’ class, it also contains field 'field', which stores interpolated values and interpolation grid parameters.

‘embrio3d’ is list with following fields: three coordinates 'x3d', 'y3d', 'z3d' and gene activity 'value'.

Examples

file <- system.file("extdata/data", "ab16.txt", package = "BioSSA")
df <- read.emb.data(file)

emb2d <- embryo2d(cad ~ AP + DV, data = df)
plot(emb2d, "nuclei-2d")

[Package BioSSA version 0.0.4 Index]