plot.embryo3d {BioSSA}	R Documentation

Plot ‘embryo3d’ object

Description

plot method for class ‘embryo3d’

Usage

## S3 method for class 'embryo3d'
plot(x, type = c("hull", "nuclei"), ...)

Arguments

x
object of class ‘embryo3d’

type
plot type

...
further arguments passed to plot functions

Details

For type = 'hull', the convex hull of nuclei is depicted, the color is determined by intensities. At edges the color is obtained by linear interpolation.

For type = 'nuclei' the nuclei are deposted as color spheres. Color is determined by intensities. Warning: Drawing can take a long time if the number of nuclei is large.

Parameter col is vector of two colors corresponding to minimal and maximal intensivities. Default value is c('blue', 'red').

Other parameters are passed to functions shade3d and plot3d correspondingly.

Examples

file <- system.file("extdata/pce", "v5_s12166-23au07-22.pce", package = "BioSSA")
pce <- read.pce.gene(file, gene = "eve")
plot(pce, type = "hull")
plot(pce, type = "nuclei")

[Package BioSSA version 0.1 Index]