reconstruct {BioSSA}	R Documentation

Component reconstruction for BioSSA decomposition object

Description

Reconstruct the data given the BioSSA decomposition and the desired grouping of the elementary components.

Usage

## S3 method for class 'BioSSA2d'
reconstruct(x, groups, ...)
## S3 method for class 'BioSSA3d'
reconstruct(x, groups, ...)

Arguments

x
BioSSA decomposition object

groups
list of numeric vectors, indices of elementary components used for reconstruction, the entries of the list can be named, see 'Value' for more information

...
further arguments passed to corresponding SSA reconstruction method

Details

For ‘embryo2d’ object, the function run reconstruct method for stored ‘2d.ssa’ object and then performs back interpolation to nuclei positions.

For ‘embryo3d’ object, the function run reconstruction for all ‘embryo2d’ that are stored within ‘embrio3d’, then perform desweeping and joining reconstructed fragments of whole 3d embryo.

Value

List of reconstructed embryos (‘embrio2d’ or ‘embrio3d’ correspondingly). Elements of the list have the same names as elements of groups. If the group is unnamed, then corresponding component will obtain name ‘Fn’, where ‘n’ is its index in groups list.

See Also

BioSSA, reconstruct.

[Package BioSSA version 0.0.4 Index]