BioSSA-class {BioSSA}	R Documentation

S3 classes for BioSSA decompositions

Description

Class for holding BioSSA decomposition

Details

‘BioSSA2d’ contains 2d data, their interpolation into a regular grid and the deomposition of the interpolated data (of their trajectory matrix).

‘BioSSA3d’ contains 3d data and the list of its planar projections in the form of ‘BioSSA2s’ class (at the moment, it consists of only one cylindrical projection).

See Also

BioSSA.

Examples

xlim <- c(22, 88)
ylim <- c(32, 68)
L <- c(15, 15)

file <- system.file("extdata/data", "ab16.txt", package = "BioSSA")
df <- read.emb.data(file)

bss <- BioSSA(cad ~ AP + DV, data = df, ylim = ylim, xlim = xlim, L = L)

[Package BioSSA version 0.1 Index]