BioSSA-package {BioSSA}	R Documentation

A collection of methods for singular spectrum analysis of drosophila embryo (2d and 3d) gene expression

Description

This package provides routines for SSA analysis of gene expression.

Details

This package is designed for analysis of drosophila embryo gene expression given in 2d or 3d form. Data should contain intensities of gene expression and coordinates, 2d or 3d correspondingly.

The main function is BioSSA, which performs singular spectrum analysis for further decomposition on pattern and noise. The data should be prepared as a data frame and the user can set the names of columns contining the data. After BioSSA is applied, the reconstruction of the image components can be obtained by the reconstruct function. Analysis of residuals is performed by noise.model methods.

S3-classes ‘BioSSA’ contain decompositions of 2d and 3d embryos. S3-class ‘noise.model’ constais residuals analysis results.

Visual presentation of the decomposition and residuals' analysis results can be performed by plot-methods of corresponding classes.

Author(s)

Alex Shlemov, Nina Golyandina

Maintainer: Alex Shlemov <shlemovalex@gmail.com>

References

Golyandina N. et al (2012): Measuring Gene Expression Noise in Early Drosophila Embryos: Nucleus-to-nucleus Variability. Procedia Computer Science Volume 9, 2012, Pages 373–382.

Golyandina N., Korobeynikov A., Shlemov A. and Usevich K. (2013): Multivariate and 2D Extensions of Singular Spectrum Analysis with the Rssa Package http://arxiv.org/abs/1309.5050

Shlemov A. and Golyandina N. (2014): Shaped extensions of singular spectrum analysis http://arxiv.org/abs/1401.4980

See Also

BioSSA, noise.model.

Examples

bad <- 1
good <- 3
wave <- 3
xlim <- c(22, 88)
ylim <- c(32, 68)
atx <- 50
aty <- 50
tolx <- 5
toly <- 5
L <- c(15, 15)

file <- system.file("extdata/data", "ab16.txt", package = "BioSSA")
df <- read.emb.data(file)

bss <- BioSSA(cad ~ AP + DV, data = df, ylim = ylim, xlim = xlim, L = L)

## Error in .setupMethodsTables(fdef, initialize = TRUE): trying to get slot "group" from an object of a basic class ("NULL") with no slots

bss <- BioSSA(cad ~ AP + DV,  data = df,
              L = L,
              step = 0.5,
              xlim = xlim, ylim = ylim)

## Error in .setupMethodsTables(fdef, initialize = TRUE): trying to get slot "group" from an object of a basic class ("NULL") with no slots

# w-correlations for identification
plot(plot(bss, type = "ssa-wcor", groups = 1:30))

## Error in plot(bss, type = "ssa-wcor", groups = 1:30): object 'bss' not found

# Reconstruction of elementary components
rec.elem <- reconstruct(bss, groups = 1:6)

## Error in reconstruct(bss, groups = 1:6): object 'bss' not found

plot(plot(rec.elem))

## Error in plot(rec.elem): object 'rec.elem' not found

# Sections for testing the reconstruction quality
rec <- reconstruct(bss, groups = list(good = 1:good, bad = 1:bad))

## Error in reconstruct(bss, groups = list(good = 1:good, bad = 1:bad)): object 'bss' not found

p.ny <- plot(attr(rec, "series"), type = "nuclei-section", at = aty, coord = "y", tol = toly)

## Error in plot(attr(rec, "series"), type = "nuclei-section", at = aty, : object 'rec' not found

p.fy1 <- plot(rec$bad, type = "field-section", at = aty, coord = "y")

## Error in plot(rec$bad, type = "field-section", at = aty, coord = "y"): object 'rec' not found

p.fy2 <- plot(rec$good, type = "field-section", at = aty, coord = "y")

## Error in plot(rec$good, type = "field-section", at = aty, coord = "y"): object 'rec' not found

p.nx <- plot(attr(rec, "series"), type = "nuclei-section", at = atx, coord = "x", tol = tolx)

## Error in plot(attr(rec, "series"), type = "nuclei-section", at = atx, : object 'rec' not found

p.fx1 <- plot(rec$bad, type = "field-section", at = atx, coord = "x")

## Error in plot(rec$bad, type = "field-section", at = atx, coord = "x"): object 'rec' not found

p.fx2 <- plot(rec$good, type = "field-section", at = atx, coord = "x")

## Error in plot(rec$good, type = "field-section", at = atx, coord = "x"): object 'rec' not found

# y-sections, bad and good
pls <- list()
pls[[1]] <- p.ny + p.fy1

## Error in eval(expr, envir, enclos): object 'p.ny' not found

pls[[2]] <- plot(residuals(bss, 1:bad), type = "nuclei-section",
                 at = aty, coord = "y", tol = toly,
                 ref = TRUE, col = "blue")

## Error in residuals(bss, 1:bad): object 'bss' not found

pls[[3]] <- p.ny + p.fy2

## Error in eval(expr, envir, enclos): object 'p.ny' not found

pls[[4]] <- plot(residuals(bss, 1:good), type = "nuclei-section",
                 at = aty, coord = "y", tol = toly,
                 ref = TRUE, col = "blue")

## Error in residuals(bss, 1:good): object 'bss' not found

print(pls[[1]], split = c(1, 1, 2, 2), more = TRUE)

## Error in pls[[1]]: subscript out of bounds

print(pls[[2]], split = c(2, 1, 2, 2), more = TRUE)

## Error in pls[[2]]: subscript out of bounds

print(pls[[3]], split = c(1, 2, 2, 2), more = TRUE)

## Error in pls[[3]]: subscript out of bounds

print(pls[[4]], split = c(2, 2, 2, 2))

## Error in pls[[4]]: subscript out of bounds

# x-sections, bad and good
pls <- list()
pls[[1]] <- p.nx + p.fx1

## Error in eval(expr, envir, enclos): object 'p.nx' not found

pls[[2]] <- plot(residuals(bss, 1:bad), type = "nuclei-section",
                 at = atx, coord = "x", tol = tolx,
                 ref = TRUE, col = "blue")

## Error in residuals(bss, 1:bad): object 'bss' not found

pls[[3]] <- p.nx + p.fx2

## Error in eval(expr, envir, enclos): object 'p.nx' not found

pls[[4]] <- plot(residuals(bss, 1:good), type = "nuclei-section",
                 at = atx, coord = "x", tol = tolx,
                 ref = TRUE, col = "blue")

## Error in residuals(bss, 1:good): object 'bss' not found

print(pls[[1]], split = c(1, 1, 2, 2), more = TRUE)

## Error in pls[[1]]: subscript out of bounds

print(pls[[2]], split = c(2, 1, 2, 2), more = TRUE)

## Error in pls[[2]]: subscript out of bounds

print(pls[[3]], split = c(1, 2, 2, 2), more = TRUE)

## Error in pls[[3]]: subscript out of bounds

print(pls[[4]], split = c(2, 2, 2, 2))

## Error in pls[[4]]: subscript out of bounds

# Dependence of noise on trend
p1 <- plot(bss, type = "residuals", model = "additive", groups = 1:good)

## Error in plot(bss, type = "residuals", model = "additive", groups = 1:good): object 'bss' not found

p2 <- plot(bss, type = "residuals", model = "poisson", groups = 1:good)

## Error in plot(bss, type = "residuals", model = "poisson", groups = 1:good): object 'bss' not found

p3 <- plot(bss, type = "residuals", model = "multiplicative", groups = 1:good)

## Error in plot(bss, type = "residuals", model = "multiplicative", groups = 1:good): object 'bss' not found

print(p1, split = c(1, 1, 3, 1), more = TRUE);

## Error in print(p1, split = c(1, 1, 3, 1), more = TRUE): object 'p1' not found

print(p2, split = c(2, 1, 3, 1), more = TRUE);

## Error in print(p2, split = c(2, 1, 3, 1), more = TRUE): object 'p2' not found

print(p3, split = c(3, 1, 3, 1));

## Error in print(p3, split = c(3, 1, 3, 1)): object 'p3' not found

# 3d-figure of reconstruction
plot(rec$good, type = "nuclei-3d", col = c("blue", "red"))

## Error in plot(rec$good, type = "nuclei-3d", col = c("blue", "red")): object 'rec' not found

# 2d-figures with triangulation
plot(rec$good, type = "nuclei-2d")

## Error in plot(rec$good, type = "nuclei-2d"): object 'rec' not found

plot(rec.elem[[wave]], type = "nuclei-2d")

## Error in plot(rec.elem[[wave]], type = "nuclei-2d"): object 'rec.elem' not found

# Model estimation
noise.model(bss, 1:good)

## Error in noise.model(bss, 1:good): object 'bss' not found

[Package BioSSA version 0.2 Index]