decompose {BioSSA}	R Documentation

Performs the SSA decomposition of stored ‘2d.ssa’ object(s)

Description

Perform the SSA decomposition of stored ‘2d.ssa’ object(s).

Usage

## S3 method for class 'BioSSA2d'
decompose(x, ...)
## S3 method for class 'BioSSA3d'
decompose(x, ...)

Arguments

x
object of class ‘BioSSA2d’ or ‘BioSSA3d’

...
further arguments passed to the internal decompose method call

Details

For the ‘BioSSA2d’ object decompose calls the decompose method from ‘Rssa’ package applied to the ‘2d.ssa’ object stored in input ‘BioSSA2d’ object.

For the 'BioSSA3d' object decompose calls the decompose method applied to all stored ‘BioSSA2d’ objects.

See Also

decompose from package ‘Rssa’.

[Package BioSSA version 0.2 Index]