plot.BioSSA2d {BioSSA}	R Documentation

Plot method for class ‘BioSSA2d’

Description

Plot the decomposition of 2d embryo. Plot graphs from the ‘Rssa’ package for ‘ssa’ object. Also, it can draw two plots of residuals: plot of the corresponding ‘noise.model’ and scatterplot of residuals for different trends reconstructed by different sets of elementary components.

Usage

## S3 method for class 'BioSSA2d'
plot(x, type = c("ssa-values", "ssa-vectors", "ssa-series", "ssa-wcor",
                                      "residuals"), ...)

Arguments

x
object of class ‘BioSSA2d’

type
plot type

...
additional arguments passed to plot functions

Details

For the type parameter 'ssa-*' are just corresponding ‘Rssa’ plots.

For type = 'residuals' one can use following additional arguments:

groups

numeric vector, elementary component indices for trend extraction

groups.y

like previous, used for ‘residuals-vs-residuals’ plot

lm

logical, whether plot linear regression line (with zero offset) on ‘residuals-vs-residuals’ plot

If groups.y is missing then noise model will be estimated and plotted (additional arguments will be passed to both noise.model and plot methods). Otherwise, ‘residuals-vs-residuals’ plot will be plotted.

Value

A Trellis object.

Examples

xlim <- c(22, 88)
ylim <- c(32, 68)
L <- c(15, 15)

file <- system.file("extdata/data", "ab16.txt", package = "BioSSA")
df <- read.emb.data(file)

bss <- BioSSA(cad ~ AP + DV, data = df, ylim = ylim, xlim = xlim, L = L)

## Error in .setupMethodsTables(fdef, initialize = TRUE): trying to get slot "group" from an object of a basic class ("NULL") with no slots

plot(bss, type = "ssa-values")

## Error in plot(bss, type = "ssa-values"): object 'bss' not found

plot(bss, type = "ssa-vectors")

## Error in plot(bss, type = "ssa-vectors"): object 'bss' not found

plot(bss, type = "ssa-series", groups = 1:3)

## Error in plot(bss, type = "ssa-series", groups = 1:3): object 'bss' not found

plot(bss, type = "residuals", groups = 1:3)

## Error in plot(bss, type = "residuals", groups = 1:3): object 'bss' not found

plot(bss, type = "residuals", groups = 1:3, groups.y = 1:2)

## Error in plot(bss, type = "residuals", groups = 1:3, groups.y = 1:2): object 'bss' not found

[Package BioSSA version 0.2 Index]